requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
loguru==0.7.3
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.6.3
